Create entropy_GPT4.ino
To generate random numbers using the Atmega328p watchdog timer jitter on an Arduino Uno, you can leverage the watchdog timer and the noise introduced by the internal oscillator. Here's a simple Arduino sketch that generates random numbers using the watchdog timer jitter.

This sketch sets up the watchdog timer to trigger an interrupt every 250ms. In the interrupt service routine, the ISR(WDT_vect) function, it calculates the difference between the current time and the last time the watchdog timer fired. The least significant bit of the time difference is used as a source of randomness, which is XORed into the current random number.

The generated random numbers are printed to the serial monitor. Note that this method may not produce high-quality random numbers and should not be used for cryptographic purposes. For better random number generation, consider using a dedicated hardware random number generator or a library like Entropy.