Create mpu6050_comparator.ino
An Arduino program (written by GPT-4) where one Arduino Uno acts as a comparator to another Arduino Uno generating telemetry data from an MPU6050 3-axis gyroscope and a 3-axis accelerometer.

We'll set up the first Arduino (the sensor) to collect the data from the MPU6050 and send it to the second Arduino (the comparator) via Serial communication. This program is the comparator.

Remember to connect the TX pin of the first Arduino to the RX1 pin of the second Arduino, and also connect their GND pins together. In this example, I assumed the second Arduino is a Mega or other variant with multiple hardware serial ports. If you're using a second Uno, you'll need to use the SoftwareSerial library to create a second serial port.