Update mpu6050_comparator.ino
Ref https://github.com/ElectronicCats/mpu6050

An Arduino program (written by GPT-4) where one Arduino Uno acts as a comparator to another Arduino Uno generating telemetry data from an MPU6050 3-axis gyroscope and a 3-axis accelerometer. This is the comparator program. 

We'll set up the first Arduino (the sensor) to collect the data from the MPU6050 and send it to the second Arduino (the comparator) via Serial communication.