Create laser_impact_offset.ino
GPT-4: Creating a flight control system for a propelled canister using an Arduino Uno can be complex. However, I can provide you with a simple example that demonstrates how to receive data from an optical sensor, determine if a predetermined point has been traversed, and respond to the laser designation. Note that implementing a fully functional flight control system would require additional hardware, such as servo motors or an actuation system, as well as more advanced algorithms for stabilization and guidance.

This example assumes that you have an optical sensor that sends data to the Arduino Uno, and the data represents whether the canister is aligned with the laser designator. The optical sensor's output should be in the form of a HIGH or LOW signal, indicating whether the canister is aligned with the laser or not.

In this example, the alignWithLaser() function contains a simple logic to adjust the servo positions. You should replace this with more sophisticated control algorithms to achieve the desired accuracy and stability during the descent of the canister. Additionally, you need to implement the getCurrentHeight() and moveToOffsetIndicator() functions to handle altitude readings and canister guidance, respectively. See patent: US8766152B1