Create laser_impact_PIDALGO.ino
Combine with laser_impact_offset.ino
Ref: https://github.com/br3ttb/Arduino-PID-Library/tree/master

To implement a more advanced PID control algorithm and use data from an MPU6050 for the canister's orientation and speed, you'll need to install the required libraries and add the necessary code to your existing sketch.

First, install the following libraries using the Arduino Library Manager:

"MPU6050_tockn" - for reading data from the MPU6050
"PID_v1" - for implementing the PID control
Next, include the necessary libraries and initialize the MPU6050 and PID controllers:

Now, you can update the moveToOffsetIndicator() function to use the PID controllers and the MPU6050 data for canister guidance:

In this updated function, we calculate the desired position (setpoint) based on the canister's orientation and speed using the MPU6050 data. The PID controllers then generate control outputs to guide the canister towards the target position, considering its orientation and speed.

Please note that you'll likely need to tune the PID controller gains (kP, kI, and kD) to achieve the best performance and stability for your specific hardware and setup. You may also need to adjust the target position calculation based on the canister's orientation and speed, depending on the requirements of your application.