Create lsr_ofst_MPUPIDautotune.ino
Ref: https://github.com/br3ttb/Arduino-PID-Library/tree/master
Ref: https://github.com/ElectronicCats/mpu6050
See patent: US8766152B1

In this refactored sketch, we will use telemetry values from the MPU6050 to automatically adjust the PID tuning parameters while the system is running. The code will update the PID gains based on the magnitude of the MPU6050's gyro data. This approach is just an example and may not be suitable for every application; you may need to adapt the algorithm based on your specific requirements.

Here's the complete refactored Arduino sketch: