Update lsr_ofst_MPUPIDautotune.ino
Ref: https://github.com/br3ttb/Arduino-PID-Library/tree/master
Ref: https://github.com/ElectronicCats/mpu6050
See patent: US8766152B1

In this refactored sketch, we will use telemetry values from the MPU6050 to automatically adjust the PID tuning parameters while the system is running. The code will update the PID gains based on the magnitude of the MPU6050's gyro data. This approach is just an example and may not be suitable for every application; you may need to adapt the algorithm based on your specific requirements.

I assume that the laser's position is given by the targetX and targetY variables. The control logic will use the PID controllers to guide the canister to the target position. The inputX and inputY variables represent the current position of the canister, which are obtained from the getCurrentX() and getCurrentY() functions. The outputX and outputY variables will be the output of the PID controllers, which control the servos' position to align the canister with the laser.

This function uses the PID controllers to calculate the necessary servo position adjustments to align the canister with the target (laser) position. The servo positions are then constrained to the valid range (0 to 180 degrees) and written to the servo motors.

Here's the complete refactored Arduino sketch: