Create MEGA_laser_Flyer_psedcode.txt
Write an Arduino Mega flight control system in arduino-type pseudocode that guides an in-flight canister via detection of an optically illuminated target using two servo motors (for control control vanes). Use an Adafruit Ultimate GPS Breakout to determine the canister's location upon launch, and an Adafruit BME280 Temperature, Humidity, and Barometric Pressure Sensor to measure the altitude of the flying canister. Integrate streaming telemetry data from an MPU6050 3-axis gyroscope and a 3-axis accelerometer with the GPS breakout board and BME280 to determine the canister's orientation and speed when guiding it to the target. Utilize an advanced control algorithm PID (Proportional-Integral-Derivative) to achieve better performance and stability. Be sure that the PID has the ability to change tuning parameters while the system is running (in flight). See this website for reference: http://brettbeauregard.com/blog/2011/04/improving-the-beginner’s-pid-tuning-changes/