Create MEGA_Laser_guide2target.ino
The pseudocode provided does not include a complete control algorithm to home in on the laser-illuminated target during the descent of an arcing trajectory for the canister. The alignWithLaser() function only demonstrates how to align the canister with the target using PID controllers based on the current position of the canister.

To make the system home in on the laser-illuminated target during the descent, you would need to integrate the telemetry data from the GPS, BME280, and MPU6050 sensors, along with the optical sensor input, to implement a more sophisticated control algorithm that accounts for the canister's trajectory, altitude, and speed.

Please refactor the arduino Mega code in order to make the system home in on the laser-illuminated target during it's in-flight descent by integrating telemetry data from the GPS, BME280, and MPU6050 sensors, along with the optical sensor input. Also implement a feedback mechanism that uses control vanes, controlled by servo motors, to adjust the canister's flight path to guide it towards the target. 

Here's a refactored Arduino Mega flight control system sketch that uses the telemetry data from the GPS, BME280, and MPU6050 sensors, along with the optical sensor input, to home in on the laser-illuminated target during the in-flight descent. It also includes a feedback mechanism that uses control vanes, controlled by servo motors, to adjust the canister's flight path to guide it towards the target.

Please note that this sketch assumes that the getCurrentX() and getCurrentY() functions have been implemented to calculate the canister's current position based on telemetry data from the sensors. The control algorithm in the guideCanisterToTarget() function can be further optimized and tuned for the specific application.

You would also need to design a feedback mechanism that uses control vanes, controlled by servo motors, to adjust the canister's flight path to guide it towards the target. The provided pseudocode does not include this specific control logic, but it provides a foundation for integrating the required sensors and control algorithm.

This refactored sketch includes a guideCanisterToTarget() function that integrates the telemetry data from the GPS, BME280, and MPU6050 sensors, along with the optical sensor input, to guide the canister towards the laser-illuminated target during its in-flight descent. The PID controllers are used to adjust the servo motor positions, which in turn control the control vanes to change the canister's flight path.