Create 328p_diagnostics.ino
The Arduino Uno has an ATmega328P microcontroller that supports a variety of features, such as:

GPIO (General Purpose Input Output): These pins can be tested by connecting them together in pairs and then running a program that sends a signal out one pin and checks if it comes back in on the other.

Analog to Digital Converter (ADC): These can be tested by connecting them to known voltage levels and checking that the correct value is read.

Serial (USART): This can be tested with loopback (connecting TX to RX) and sending data to see if it is correctly received.

Watchdog Timer: This can be tested by setting it up to reset the microcontroller and then causing a situation where it should be triggered.

Below is an example program that demonstrates how these tests might be done. This code is just a basic outline and may need further customization to meet your specific needs.

This code sets up a basic test for the GPIO pins, the ADC, the serial port (with loopback), and the watchdog timer. The GPIO test writes a HIGH and then a LOW signal to one pin and checks that the signal is correctly received on the other pin. The ADC test simply reads the value from A0 and prints it. The serial test writes a 'T' and checks that it is received (you need to connect RX to TX for this to work). The watchdog timer test simply waits for longer than the watchdog timeout and expects the system to reset.

Please make sure to use this code responsibly, since it involves the use of a watchdog timer that will reset the system. Be careful when testing this functionality, as it could potentially cause issues with your program if not used correctly.