Create 328p_Ffourierv4Rand.ino
One way to increase entropy is to gather more random data. We can do this by calling Entropy.random() multiple times and then combining the results.

The function generateRandom() generates a random number with higher entropy by combining 3 calls to Entropy.random() and averaging them to get a number in the correct range. We then replace Entropy.random(SAMPLES) with generateRandom() in the loop function."

The generateRandom() function generates a more entropy-rich random number by averaging the results of three calls to Entropy.random(SAMPLES). This newly generated number is used as the input to the FFT computation, potentially making the results more diverse and less predictable, and thus increasing the robustness of the FFT speed test.

In the loop function, this change ensures that each real part of the input data is generated with a higher entropy random number.

It's important to note, however, that while this method does indeed increase the randomness of the input data, it may also increase the execution time of the random number generation. Be sure to take this into account when interpreting the results of the FFT speed test.