Create Create 328p_Ffv5RandSP.ino
Can we refactor the previous code to include elements/features which will make it more visually intuitive via the serial plotter window?

The Serial Plotter built into the Arduino IDE is a great tool for visualizing numerical data over time. You could modify the sketch to plot the real part of the FFT input data and the magnitude of the FFT output data, to provide a more intuitive view of what's happening.

Here's a version of the sketch that sends the input and output data to the Serial Plotter:

n this updated code, we've added Serial.println(vReal[i]); in two places:

After generating the random input data: this will plot the real part of the FFT input data.

After computing the FFT and converting the complex FFT output to magnitude: this will plot the magnitude of the FFT output data.

This will give you a time-domain plot of the generated input data and a frequency-domain plot of the FFT output data. Please note that since both inputs and outputs are being plotted in the same plot, their scales will be off. To distinguish between input and output, you may consider introducing a delay or marker between the input and output plots. Also remember to keep the Serial Plotter open while the sketch is running.