Create Create 328p_Ffv6RandSP.ino
Since both inputs and outputs are being plotted in the same plot, their scales are off. To distinguish between input and output, let's refactor by introducing a delay or marker between the input and output plots.

we can introduce a unique marker to distinguish between the input and output plots. In this case, we could use -1 as a marker, since we're only expecting positive numbers in both our input and output data. We can print this marker to the Serial monitor right after finishing with the input data and right before starting with the output data. We can also add a short delay for better visibility.

Here's the code with the markers added:

With this updated version, you'll see the input data, followed by "-1" (as a marker), followed by the output data in your Serial Plotter. The small delay helps in better visibility of these transitions.