Create Entropy_Alphanumstring.ino
Using Walter Anderson's Entropy.h library To generate random alphanumeric strings. 

Even words, phrases, or sentences based on the random numbers generated by the Entropy library, we need to map the generated numbers to a character set or a dictionary.

For alphanumeric characters, we can use ASCII values. For words, phrases, or sentences, we'll need a dictionary of words, which can be large and memory-consuming, especially for a microcontroller like the Arduino. One way around this is to use a predefined set of words, or use a pseudoword generator to create pronounceable, but not necessarily meaningful, English-like words.

Below is an example of mapping the random numbers to alphanumeric characters:

As for generating words, phrases, or sentences, this can be a bit tricky, especially when dealing with syntax and meaning. A pseudoword generator could help to create English-like words, but for generating phrases or sentences that are grammatically correct, we'll need a more complex language model, possibly one that is trained on a large dataset of English text. An Arduino would not be capable of running such a model due to memory and processing power limitations.

However, if you are working with a more powerful board or platform that can handle such tasks, there are libraries like TensorFlow Lite for Microcontrollers that can run small machine learning models on microcontrollers. You might also consider transmitting the random numbers to a more powerful system (like a PC or server) to do the text generation, then sending the results back to the Arduino for further processing or display.

Let's modify the previous code such that for each random number generated we print the random number and then alphanumeric character mapped string. 

To achieve this, you can store the alphanumeric characters in a string and print the string together with the random number. Here is how you could modify the code:

In this code, we generate a string of 23 alphanumeric characters. You can adjust the length of the string by changing the number of iterations in the for loop. Each iteration generates one random alphanumeric character. The generated alphanumeric string is then printed together with the random number.