Create Entropy_AlphanumPseduwrd.ino
a pseudoword generator to create pronounceable, but not necessarily meaningful, English-like words. These pseudowords should be output in addition to the alphanumeric strings.

Generating pseudowords can be a complex task, especially if we aim to ensure pronounceability and some semblance to English word structure. Here's a simple algorithm, though, that could work reasonably well:

Choose a random length for the word (e.g., between 2 and 10 letters).
Start with an empty word.
Add a random consonant.
With a 50% chance, add a random vowel.
Repeat steps 3 and 4 until the word is of the chosen length.
This isn't a perfect method, but it can generate words that look somewhat English-like. Here's an implementation:

In this code, the function generatePseudoword() generates a pseudoword of a random length (between 2 and 10 letters) using the simple algorithm outlined above. The pseudoword is then printed alongside the random number and alphanumeric string. Note that this simple pseudoword generator may not always create very English-like or pronounceable words, but it's a starting point. You could modify the algorithm and data to better suit your specific needs.