Create 328p_mandlebrot.ino
Rendering the Mandelbrot set on an Arduino can be a challenging task due to the resource limitations of the platform. However, it's certainly possible to do so, especially with the help of an external display module such as the Adafruit ILI9341 LCD display and the corresponding Adafruit_ILI9341 library.

Assuming you have such a display, here is a simple Arduino program that will render the Mandelbrot set:

In this program, mandelbrot(real, imag) function calculates the value of each point in the Mandelbrot set, while the colorMap(value) function maps the Mandelbrot value to a color that can be displayed on the ILI9341 LCD display.

The main drawMandelbrot() function loops over each pixel on the display, calculates the corresponding point in the Mandelbrot set, and colors the pixel accordingly. This function is called once in the main loop, after which the program stops executing to let you admire the result.

Note: Be sure to connect the ILI9341 display to the correct pins as defined in the program, or adjust the pin definitions as needed to match your hardware setup.

Also, remember that rendering the Mandelbrot set can be a slow process, particularly on a resource-limited platform like the Arduino. It might take several minutes to render the entire set, depending on the complexity of the fractal and the speed of your hardware.