Create 328p_mandelbrot_serial.ino
Creating a text-based rendering of the Mandelbrot set in the Arduino IDE's serial monitor is certainly possible, but the results will be quite crude due to the limited resolution and lack of grayscale or color.

Here is a simple program that creates a 32x16 ASCII art representation of the Mandelbrot set:

This sketch will output an ASCII representation of the Mandelbrot set to the serial monitor when it starts. The mandelbrot(real, imag) function is used to calculate the value of each point in the Mandelbrot set. Points that reach the maximum number of iterations before escaping are considered part of the set and are represented with a #, while all other points are represented with a space.

Remember that the Arduino's Serial Monitor should be set to "No line ending" and the baud rate should match the one specified in the begin() method of the Serial object (9600 in this case).

Please note that due to the limited resolution and inability to display grayscale or color in the Serial Monitor, the output will be a very crude and blocky representation of the Mandelbrot set.