Create 328p_mandelbrot_hres.ino
The "resolution" in this case is determined by the character grid of the Arduino Serial Monitor, which can't be changed. But there are a few tricks to get a bit more detail out of our ASCII rendering:

Use more ASCII characters: Rather than just using a space and '#', we can use different characters to represent different levels of iteration count before escape. For example, '.','-','+','*','#' can be used to give a crude form of grayscale.

Use a larger grid: If you don't mind scrolling around in the Serial Monitor, you can use a larger grid for your rendering. This will take more computation time, but will give you a more detailed picture.

Here's a simple modification of the previous code that uses both of these tricks:

This sketch uses a 64x32 character grid and 5 different characters to represent the Mandelbrot set. Note that because of the increased resolution and grayscale, this sketch will take longer to run and may output a lot of data to the Serial Monitor. Be sure to set your Serial Monitor to a high enough baud rate (9600 is used in the sketch) to keep up with the data output.

This sketch now compiles and runs correctly. It will output an ASCII representation of the Mandelbrot set to the serial monitor when it starts, with a crude form of grayscale representation based on the number of iterations taken by each point to escape. The greater the number of iterations, the "darker" the point will appear.